requirements.txt
python-dotenv==1.0.1
Django==5.1.3 
psycopg2-binary==2.9.10
djangorestframework==3.15.2